Had to split the URLUtil since another logic class used URLUtil functionality
